Move data info badge into components to prepare for being used in ma progression page
