+++ frontend/src/views/DashboardManager/DiversificationCard.vue
@@ -30,7 +30,7 @@
 
 <script>
 import Constants from "@/constants"
-import DataInfoBadge from "./DataInfoBadge"
+import DataInfoBadge from "@/components/DataInfoBadge"
 import keyMeasures from "@/data/key-measures.json"
 import { hasStartedMeasureTunnel, applicableDiagnosticRules } from "@/utils"
